Fix: allow enable TLS with just server cert (#266)
When enabling TLS support, only the server certificate is required. The
client certificate and key are optional. Sometimes you would like to enable
TLS for the database but without using a client verification. You only want
TLS and only validate the server using a CA certificate. Another small change
is that you could also specify the server and client certificates and the
client key defining the PEM in one line.